Des avancées

+++ baseRoulante/AVR/asservissement_2/lib/Cos.h
@@ -1,8 +1,10 @@
 #include "TableCos.h"
 #include <avr/pgmspace.h>
 #include <stdint.h> 
+#include "serial.h"
+
 #define ABS(x) ( (x)>0?(x):-(x))
 
-extern long getCos(float angleRadian);
+extern int getCos(float angleRadian);
 
 extern float getSin(float angleRadian);
